wip: generating cycles

+++ database/migrations/2023_09_02_135746_create_cycles_table.php
@@ -14,7 +14,7 @@ return new class () extends Migration {
 
             $table->date('start');
             $table->date('end');
-            $table->integer('training_max');
+            $table->decimal('training_max');
             $table->mediumText('notes')->nullable();
 
             $table->softDeletes();
